refactor: Profile page component 제거
