+++ src/router/index.js
@@ -3,7 +3,6 @@ import Router from 'vue-router';
 import Main from '@/components/pages/Main';
 import Dashboard from '@/components/pages/Dashboard';
 import Register from '@/components/pages/Register';
-import Profile from '@/components/pages/Profile';
 import Application from '@/components/pages/Application';
 import NotFoundError from '@/components/pages/404';
 import InternalError from '@/components/pages/500';
@@ -28,7 +27,7 @@ export default new Router({
     {
       path: '/profile',
       name: 'Profile',
-      component: Profile,
+      component: Register,
       meta: { auth: true },
     },
     {
